+++ js/submitcpbeacon2.js
@@ -6,7 +6,6 @@ function SubmitCPBeacon2FormData() {
       $.post("submit_cpbeacon2.php", { cheibeding: cheibeding },
          function(data) {
                $('#actions_output').html(data);
-               $('#cpbeacon2form')[0].reset();
                $('#cpbeacon2form-1')[0].reset();
                $('#cpbeacon2form-2')[0].reset();
                $('#cpbeacon2form-3')[0].reset();
@@ -15,8 +14,6 @@ function SubmitCPBeacon2FormData() {
                $('#cpbeacon2form-6')[0].reset();
                $('#cpbeacon2form-7')[0].reset();
                $('#cpbeacon2form-8')[0].reset();
-               $('#cpbeacon2form-9')[0].reset();
-               $('#cpbeacon2form-10')[0].reset();
                
          });
 }