+++ js/submitcpbeacon3.js
@@ -6,7 +6,6 @@ function SubmitCPBeacon3FormData() {
       $.post("submit_cpbeacon3.php", { cheibeding: cheibeding },
          function(data) {
                $('#actions_output').html(data);
-               $('#cpbeacon3form')[0].reset();
                $('#cpbeacon3form-1')[0].reset();
                $('#cpbeacon3form-2')[0].reset();
                $('#cpbeacon3form-3')[0].reset();
@@ -15,8 +14,6 @@ function SubmitCPBeacon3FormData() {
                $('#cpbeacon3form-6')[0].reset();
                $('#cpbeacon3form-7')[0].reset();
                $('#cpbeacon3form-8')[0].reset();
-               $('#cpbeacon3form-9')[0].reset();
-               $('#cpbeacon3form-10')[0].reset();
                
          });
 }