+++ submit_opacity.php
@@ -1,9 +1,6 @@
 <?php
-
-// das grift dr input vom js script im CP ab:
-    $usgab = $_POST['opacityValue'];
-    echo $usgab;
-    
-    shell_exec("sudo /var/www/sync/dbuscontrol.sh setalpha $usgab");
-
+// this gets the input from JS:
+    $opacity = $_POST['opacityValue'];
+    echo $opacity;
+    shell_exec("sudo /var/www/sync/dbuscontrol.sh setalpha $opacity");
 ?>